requirements.txt
Flask==3.1.1
Pillow==11.3.0
openpyxl==3.1.2
python-docx==1.1.2
reportlab==4.2.2
